+++ tests/testthat/test_boot.R
@@ -80,3 +80,10 @@ test_that('printing', {
   expect_output(print(bootstraps(iris)))
 })
 
+
+test_that('rsplit labels', {
+  rs <- bootstraps(iris)
+  all_labs <- map_df(rs$splits, labels)
+  original_id <- rs[, grepl("^id", names(rs))]
+  expect_equal(all_labs, original_id)
+})