+++ tests/testthat/test_group.R
@@ -90,3 +90,10 @@ test_that('printing', {
 })
 
 
+test_that('rsplit labels', {
+  rs <- group_vfold_cv(iris, "Species")
+  all_labs <- map_df(rs$splits, labels)
+  original_id <- rs[, grepl("^id", names(rs))]
+  expect_equal(all_labs, original_id)
+})
+