+++ tests/testthat/test_mc.R
@@ -76,3 +76,12 @@ test_that('bad args', {
 test_that('printing', {
   expect_output(print(mc_cv(iris)))
 })
+
+
+test_that('rsplit labels', {
+  rs <- mc_cv(mtcars)
+  all_labs <- map_df(rs$splits, labels)
+  original_id <- rs[, grepl("^id", names(rs))]
+  expect_equal(all_labs, original_id)
+})
+